Add TargetIDConsolidatedTargetKeyDictionary extensions file
