+++ test/fixtures/generator/bwx_spec.json
@@ -358,6 +358,7 @@
                     "tools/generator/test/SetTargetConfigurationsTests.swift",
                     "tools/generator/test/SetTargetDependenciesTests.swift",
                     "tools/generator/test/Target+Testing.swift",
+                    "tools/generator/test/TargetIDConsolidatedTargetKeyDictionary+ExtensionsTests.swift",
                     "tools/generator/test/TargetIDTargetDictionary+ExtensionsTests.swift",
                     "tools/generator/test/XCScheme+ExtensionsTests.swift",
                     "tools/generator/test/XCSchemeEnvironmentVariablesExtensionsTests.swift",
@@ -683,6 +684,7 @@
                     "tools/generator/src/Extensions/Platform+Extensions.swift",
                     "tools/generator/src/Extensions/Sorting.swift",
                     "tools/generator/src/Extensions/String+Extensions.swift",
+                    "tools/generator/src/Extensions/TargetIDConsolidatedTargetKeyDictionary+Extensions.swift",
                     "tools/generator/src/Extensions/TargetIDTargetDictionary+Extensions.swift",
                     "tools/generator/src/Extensions/XCScheme+Extensions.swift",
                     "tools/generator/src/Extensions/XCSchemeEnvrionmentVariables+Extensions.swift",
