+++ tools/generator/src/Extensions/TargetIDTargetDictionary+Extensions.swift
@@ -52,33 +52,3 @@ extension Dictionary where Key == TargetID, Value == Target {
         return try firstTargetID(under: newStartIDs, where: predicate)
     }
 }
-
-// MARK: `collectPlatformsByKey`
-
-// TODO(chuck): MOVE ME!
-
-extension Dictionary where Key == TargetID, Value == ConsolidatedTarget.Key {
-    func collectPlatformsByKey(
-        targets: [TargetID: Target]
-    ) throws -> [ConsolidatedTarget.Key: Set<Platform>] {
-        var result = [ConsolidatedTarget.Key: Set<Platform>]()
-        for (targetID, key) in self {
-            let target = try targets.value(for: targetID)
-            result[key, default: []].insert(target.platform)
-        }
-        return result
-    }
-}
-
-// extension Dictionary where Key == TargetID, Value == Target {
-//     func collectPlatformsByKey(
-//         consolidatedTargetKeys: [TargetID: ConsolidatedTarget.Key]
-//     ) throws -> [ConsolidatedTarget.Key: Set<Platform>] {
-//         var result = [ConsolidatedTarget.Key: Set<Platform>]()
-//         for (targetID, target) in self {
-//             let key = try consolidatedTargetKeys.value(for: targetID)
-//             result[key, default: []].insert(target.platform)
-//         }
-//         return result
-//     }
-// }
