+++ tools/generator/test/TargetIDConsolidatedTargetKeyDictionary+ExtensionsTests.swift
@@ -1,0 +1,10 @@
+import XCTest
+
+@testable import generator
+
+// swiftlint:disable:next type_name
+class TargetIDConsolidatedTargetKeyDictionaryExtensionsTests: XCTestCase {
+    func test_collectPlatformsByKey() throws {
+        XCTFail("IMPLEMENT ME!")
+    }
+}
